% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Graph a country's population over time}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{Country name}
}
\value{
a graph of the input country's population from 1950 to 2020
}
\description{
Uses the WorldPopulation data set to find country and graph its population
from 1950 to 2020. Will produce an error if country name not found.
}
\examples{
CountryPopulation('Iraq')
CountryPopulation('United States of America')
}
