% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup country, year, matches, total attendance, and average attendance}
\format{
a data frame with 22 observations of 4 variables
\describe{
   \item{WorldCup}{The year and country name of where the World Cup was held.}
   \item{Matches}{The number of matches played.}
   \item{Totalattendance}{The total attendance for all matches.}
   \item{Averageattendance}{The average attendance per match.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
World Cup country, year, matches, total attendance, and average attendance
}
\keyword{datasets}
