% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Country populations for years 1950 to 2020}
\format{
a data frame with 231 observations of 72 variables
\describe{
   \item{CountryName}{The name of the country.}
   \item{1950 - 2020}{The population for the year in thousands.}
}
}
\usage{
WorldPopulation
}
\description{
Country populations for years 1950 to 2020
}
\keyword{datasets}
